Move timeseries plotting tests to their source test file
Update figure test hashes
