+++ sunpy/tests/figure_hashes_mpl_352_ft_261_astropy_51_animators_100.json
@@ -50,15 +50,14 @@
   "sunpy.map.tests.test_reproject_to.test_reproject_to_hpc_exact": "93171456dfac771f53bc0465922cf5811048df0b56100d329a55f71b36bdc8f5",
   "sunpy.map.tests.test_reproject_to.test_reproject_to_hpc_adaptive": "91b046ff2805fd47840c128b8b510f290cdd3cafd04b3dd283c2478a61a6a244",
   "sunpy.net.tests.test_helioviewer.test_download_jp2_map": "a71654a4a7ceea689869c68f4cd7f0ebfc5e970caed73f2e73e677745c99e0e0",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_eve_peek": "2a22f93af44971411800e4fcd0d93d2128fe2b9e6b97e576e8372c2cd12d7d78",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_esp_peek": "decd0819a6e81cfbcc708188785ee7efcf312cd07294bc36fcf3cbbe32512030",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_fermi_gbm_peek": "83f0b53ca6f1cd809bd03e4aff80b3c02c8078e552b0c3e90b2cb3cf26932a51",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_norh_peek": "44c192d4cf0f260b5dc886b82a9d5cc51db595d3cdcea834059ef9902eb3849c",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_goes_peek": "a38555a43c12e52ebcf54817bbadb7b080c693b6676bf5c07338beadcbacac7f",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_lyra_peek": "cfa100a127a39eed547d5ad5c36d2d5aa3873e0b85a208b73aa8f81caa52c459",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_rhessi_peek": "ee698fae54620b7cdef08aad2b947d98c03b318c9e4ea65729e39d81fdc93d02",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_json_ind_peek": "1bea3a56178db101abca168bc51e127b7b96f9eed58d74e3d19b6ffc14643e80",
-  "sunpy.timeseries.tests.test_timeseriesbase.test_noaa_json_pre_peek": "f1493e2c5263615795dc3ca774365dcc7c0ac1aa180e1dd60e1fbb4269360806",
+  "sunpy.timeseries.sources.tests.test_eve.test_eve_peek": "2a22f93af44971411800e4fcd0d93d2128fe2b9e6b97e576e8372c2cd12d7d78",
+  "sunpy.timeseries.sources.tests.test_fermi_gbm.test_fermi_gbm_peek": "2b1043c62d9a95c663e738c3557644cba112587d19f713a0b5c05138c101f440",
+  "sunpy.timeseries.sources.tests.test_goes.test_goes_peek": "fe27add6ae15b3c21228ff6e575982b92ba2574cf0b4accb3e4e40b025b9247a",
+  "sunpy.timeseries.sources.tests.test_lyra.test_lyra_peek": "cfa100a127a39eed547d5ad5c36d2d5aa3873e0b85a208b73aa8f81caa52c459",
+  "sunpy.timeseries.sources.tests.test_noaa.test_noaa_json_pre_peek": "135ad8db1fa70a7bbbe7b174dec0d4d5d0af03dfaddd10ee1fc23e5afcc2a0e8",
+  "sunpy.timeseries.sources.tests.test_noaa.test_noaa_json_ind_peek": "1bea3a56178db101abca168bc51e127b7b96f9eed58d74e3d19b6ffc14643e80",
+  "sunpy.timeseries.sources.tests.test_norh.test_norh_peek": "0732b7b16f8edcc7d7a52241827b79380810a70b1b8a12b9b8dd1dbe1515c6fd",
+  "sunpy.timeseries.sources.tests.test_rhessi.test_rhessi_peek": "0d4cc06897e0e058fc8b049d3502f6fc842755ad393f0b311de69e6c52a902b3",
   "sunpy.timeseries.tests.test_timeseriesbase.test_generic_ts_peek": "ba0469ba921c3367384c04de4e64d16cacd1499277a14f1809c087d8f21ee041",
   "sunpy.timeseries.tests.test_timeseriesbase.test_column_subset_peek": "0c1ac6714d9bc505db3ee449fa9a8ba6842b0d756a8c8f02a2133547123c2a92",
   "sunpy.visualization.animator.tests.test_basefuncanimator.test_lineanimator_figure": "6865a446680cb7088d71becd653e3c9e9e8e982d71956c95f690d52365bfe106",
