+++ sunpy/timeseries/sources/tests/test_goes.py
@@ -2,23 +2,19 @@ import pytest
 
 import sunpy.timeseries
 from sunpy.data.test import get_test_filepath
+from sunpy.tests.helpers import figure_test
 
-goes_filepath = get_test_filepath('go1520110607.fits')
 goes_filepath_com = get_test_filepath('go1520120601.fits.gz')
 new_goes15_filepath = get_test_filepath('goes_truncated_test_goes15.nc')
 new_goes17_filepath = get_test_filepath('goes_truncated_test_goes17.nc')
 
 
-def test_implicit_goes():
-    # Test a GOES TimeSeries
-    ts_goes = sunpy.timeseries.TimeSeries(goes_filepath)
-    assert isinstance(ts_goes, sunpy.timeseries.sources.goes.XRSTimeSeries)
+def test_implicit_goes(goes_test_ts):
+    assert isinstance(goes_test_ts, sunpy.timeseries.sources.goes.XRSTimeSeries)
 
 
-def test_implicit_goes_com():
-    # Test a GOES TimeSeries
-    ts_goes = sunpy.timeseries.TimeSeries(goes_filepath_com)
-    assert isinstance(ts_goes, sunpy.timeseries.sources.goes.XRSTimeSeries)
+def test_implicit_goes_com(goes_test_ts):
+    assert isinstance(goes_test_ts, sunpy.timeseries.sources.goes.XRSTimeSeries)
 
 
 def test_implicit_new_goes15():
@@ -33,10 +29,8 @@ def test_implicit_new_goes17():
     assert isinstance(ts_goes, sunpy.timeseries.sources.goes.XRSTimeSeries)
 
 
-def test_implicit_goes_satno():
-    # Test a GOES TimeSeries for satellite number
-    ts_goes = sunpy.timeseries.TimeSeries(goes_filepath)
-    assert ts_goes.observatory == 'GOES-15'
+def test_implicit_goes_satno(goes_test_ts):
+    assert goes_test_ts.observatory == 'GOES-15'
 
 
 def test_implicit_new_goes15_satno():
@@ -56,12 +50,6 @@ def test_implicit_goes_satno_missing():
     ts_goes = sunpy.timeseries.TimeSeries(new_goes17_filepath)
     del ts_goes.meta.metas[0]['id']
     assert ts_goes.observatory is None
-
-
-def test_goes():
-    # Test a GOES TimeSeries
-    ts_goes = sunpy.timeseries.TimeSeries(goes_filepath, source='XRS')
-    assert isinstance(ts_goes, sunpy.timeseries.sources.goes.XRSTimeSeries)
 
 
 def test_goes_com():
@@ -104,3 +92,14 @@ def test_goes_remote():
     goes = sunpy.timeseries.TimeSeries(
         'https://umbra.nascom.nasa.gov/goes/fits/2018/go1520180626.fits')
     assert isinstance(goes, sunpy.timeseries.sources.goes.XRSTimeSeries)
+
+
+def test_goes_plot_column(goes_test_ts):
+    ax = goes_test_ts.plot(columns=['xrsa'])
+    assert len(ax.lines) == 1
+    assert '0.5--4.0' == ax.lines[0].get_label().split()[0]
+
+
+@figure_test
+def test_goes_peek(goes_test_ts):
+    goes_test_ts.peek()
